Extract EventCalendar
